21/01

+++ Obrabook/src/main/java/com/example/Obrabook/repository/ClienteRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.Obrabook.model.Cliente;
 
 @Repository
-public interface ClienteRepository extends MongoRepository<Cliente, Integer> {
+public interface ClienteRepository extends MongoRepository<Cliente, Long> {
     
 }
